moved ProfileHint outside of internal
